+++ src/uqtestfuns/test_functions/sobol_g.py
@@ -3,42 +3,52 @@ Module with an implementation of the Sobol-G test function.
 
 The Sobol'-G function is an M-dimensional scalar-valued function.
 It was introduced in [1] for testing numerical integration algorithms
-(e.g., quasi-Monte-Carlo; see for instance [2].
+(e.g., quasi-Monte-Carlo; see also for instance [2] and [3]).
+The current form (and name) was from [4] and used in the context of global
+sensitivity analysis. There, the function was generalized by introducing
+a set of parameters that determines the importance of each input variable.
 Later on, it becomes a popular testing function for global sensitivity analysis
-methods; see, for instances, [3], [4], and [5].
+methods; see, for instances, [5], [6], and [7].
 
-The parameters of the Sobol'-G function (i.e., the coefficients) determine
-the importance of each input variable.
 There are several sets of parameters used in the literature.
 
 Notes
 -----
-- The parameters used in [3] and [4] correspond to
-  the parameter choice 3 in [1].
+- The parameters used in [5] and [6] correspond to
+  the parameter choice 3 in [3].
 
 References
 ----------
-1. I. Radović, I. M. Sobol’, and R. F. Tichy, “Quasi-Monte Carlo Methods for
+
+1. Paul Bratley, Bennet L. Fox, and Harald Niederreiter, "Implementation and
+   tests of low-discrepancy sequences," ACM Transactions on Modeling and
+   Computer Simulation, vol. 2, no. 3, pp. 195-213, 1992.
+   DOI:10.1145/146382.146385
+2. I. Radović, I. M. Sobol’, and R. F. Tichy, “Quasi-Monte Carlo Methods for
    Numerical Integration: Comparison of Different Low Discrepancy Sequences,”
    Monte Carlo Methods and Applications, vol. 2, no. 1, pp. 1–14, 1996.
    DOI: 10.1515/mcma.1996.2.1.1.
-2. I. M. Sobol’, “On quasi-Monte Carlo integrations,” Mathematics and Computers
+3. I. M. Sobol’, “On quasi-Monte Carlo integrations,” Mathematics and Computers
    in Simulation, vol. 47, no. 2–5, pp. 103–112, 1998.
    DOI: 10.1016/S0378-4754(98)00096-2
-3. A. Marrel, B. Iooss, F. Van Dorpe, and E. Volkova, “An efficient
+4. A. Saltelli and I. M. Sobol’, “About the use of rank transformation in
+   sensitivity analysis of model output,” Reliability Engineering
+   & System Safety, vol. 50, no. 3, pp. 225–239, 1995.
+   DOI: 10.1016/0951-8320(95)00099-2.
+5. A. Marrel, B. Iooss, F. Van Dorpe, and E. Volkova, “An efficient
    methodology for modeling complex computer codes with Gaussian processes,”
    Computational Statistics & Data Analysis, vol. 52, no. 10,
    pp. 4731–4744, 2008.
    DOI: 10.1016/j.csda.2008.03.026
-4. A. Marrel, B. Iooss, B. Laurent, and O. Roustant, “Calculations of Sobol
+6. A. Marrel, B. Iooss, B. Laurent, and O. Roustant, “Calculations of Sobol
    indices for the Gaussian process metamodel,” Reliability Engineering &
    System Safety, vol. 94, no. 3, pp. 742–751, 2009.
     DOI: 10.1016/j.ress.2008.07.008
-5. S. Kucherenko, B. Feil, N. Shah, and W. Mauntz, “The identification of
+7. S. Kucherenko, B. Feil, N. Shah, and W. Mauntz, “The identification of
    model effective dimensions using global sensitivity analysis,”
    Reliability Engineering & System Safety, vol. 96, no. 4, pp. 440–449, 2011.
    DOI: 10.1016/j.ress.2010.11.003
-6. T. Crestaux, J.-M. Martinez, O. Le Maître, and O. Lafitte, “Polynomial
+8. T. Crestaux, J.-M. Martinez, O. Le Maître, and O. Lafitte, “Polynomial
    chaos expansion for uncertainties quantification and sensitivity analysis,”
    presented at the Fifth International Conference on Sensitivity Analysis
    of Model Output, 2007.
@@ -96,6 +106,62 @@ AVAILABLE_INPUT_SPECS = {
 }
 
 DEFAULT_INPUT_SELECTION = "Radovic1996"
+
+
+def _get_params_saltelli_1995_1(spatial_dimension: int) -> np.ndarray:
+    """Construct a parameter array for Sobol'-G according to example 1 in [4].
+
+    Notes
+    -----
+    - The function was most probably first appear in its original form
+      in [1] (without parameters).
+    - With the selected parameters, all input variables are equally important.
+    """
+    yy = np.zeros(spatial_dimension)
+
+    return yy
+
+
+def _get_params_saltelli_1995_2(spatial_dimension: int) -> np.ndarray:
+    """Construct a parameter array for Sobol'-G according to example 2 in [4].
+
+    Notes
+    -----
+    - With the selected parameters, the first two input variables are
+      important, one is moderately important, and the rest is non-influential.
+    - Originally, the dimension is limited to 8-dimensions; if more dimensions
+      are used then the remaining dimension is also non-influential.
+    """
+    yy = np.zeros(spatial_dimension)
+    if spatial_dimension > 0:
+        yy[0] = 0
+
+    if spatial_dimension > 1:
+        yy[1] = 0
+
+    if spatial_dimension > 2:
+        yy[2] = 3
+
+    if spatial_dimension > 3:
+        yy[3:] = 9
+
+    return yy
+
+
+def _get_params_saltelli_1995_3(spatial_dimension: int) -> np.ndarray:
+    """Construct a parameter array for Sobol'-G according to example 3 in [4].
+
+    Notes
+    -----
+    - With the selected parameters, the first input variable is the most
+      important and the importance of the remaining variables is decreasing.
+    - Originally, the dimension is limited to 20-dimensions; if more dimensions
+      are used then the remaining dimension is also non-influential.
+    - The parameter set is also used in [8].
+    """
+    yy = (np.arange(1, spatial_dimension + 1) - 1) / 2.0
+
+    return yy
 
 
 def _get_params_sobol_1998_1(spatial_dimension: int) -> np.ndarray:
@@ -173,30 +239,22 @@ def _get_params_kucherenko_2011_3b(spatial_dimension: int) -> np.ndarray:
     return yy
 
 
-def _get_params_crestaux_2007(spatial_dimension: int) -> np.ndarray:
-    """Construct a parameter array for Sobol'-G according to [6]."""
-    yy = (np.arange(1, spatial_dimension + 1) - 1) / 2.0
-
-    return yy
-
-
 AVAILABLE_PARAMETERS = {
+    "Saltelli1995-1": _get_params_saltelli_1995_1,
+    "Saltelli1995-2": _get_params_saltelli_1995_2,
+    "Saltelli1995-3": _get_params_saltelli_1995_3,
     "Sobol1998-1": _get_params_sobol_1998_1,
     "Sobol1998-2": _get_params_sobol_1998_2,
     "Sobol1998-3": _get_params_sobol_1998_3,
     "Sobol1998-4": _get_params_sobol_1998_4,
     "Kucherenko2011-2a": _get_params_kucherenko_2011_2a,
     "Kucherenko2011-3b": _get_params_kucherenko_2011_3b,
-    "Crestaux2007": _get_params_crestaux_2007,
 }
 
-DEFAULT_PARAMETERS_SELECTION = "Crestaux2007"
-
-# The dimension is variable, it requires a default for fallback
-DEFAULT_DIMENSION_SELECTION = 2
+DEFAULT_PARAMETERS_SELECTION = "Saltelli1995-3"
 
 
-def evaluate(xx: np.ndarray, parameters):
+def evaluate(xx: np.ndarray, parameters: np.ndarray):
     """Evaluate the Sobol-G function on a set of input values.
 
     Parameters
@@ -204,6 +262,8 @@ def evaluate(xx: np.ndarray, parameters):
     xx : np.ndarray
         M-Dimensional input values given by an N-by-M array where
         N is the number of input values.
+    parameters : np.ndarray
+        The parameters (i.e., coefficients) of the Sobol'-G function.
 
     Returns
     -------
@@ -211,8 +271,8 @@ def evaluate(xx: np.ndarray, parameters):
         The output of the Sobol-G function evaluated on the input values.
         The output is a 1-dimensional array of length N.
     """
-    params = parameters
-    yy = np.prod(((np.abs(4 * xx - 2) + params) / (1 + params)), axis=1)
+    aa = parameters
+    yy = np.prod(((np.abs(4 * xx - 2) + aa) / (1 + aa)), axis=1)
 
     return yy
 
@@ -221,7 +281,7 @@ class SobolG(UQTestFunABC):
     """An implementation of the M-dimensional Sobol'-G test function."""
 
     _tags = ["sensitivity", "integration"]
-    _description = "Sobol'-G function from Radović et al. (1996)"
+    _description = "Sobol'-G function from Saltelli and Sobol' (1995)"
     _available_inputs = AVAILABLE_INPUT_SPECS
     _available_parameters = AVAILABLE_PARAMETERS
     _default_parameters = DEFAULT_PARAMETERS_SELECTION